+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### New Feature
 - Dynamic buffer to prevent loss of any message in case of failures
+- Added option `--home-path` to override environment variable `RESIDUE_HOME`
 
 ## [2.1.0] - 07-03-2018
 ### Fixed